refactor: split util/crane.py from util/oci.py
Crane-specific code (get_crane, BazelImage, crane_push) moves to
util/crane.py with the @crane data dep. util/oci.py retains general
OCI utilities (docker auth, digest reading, image loading) without
pulling in the crane binary.
