+++ props/testing/fixtures/e2e_infra.py
@@ -29,7 +29,8 @@ from testcontainers.core.wait_strategies import LogMessageWaitStrategy
 
 from props.db.models import AgentType
 from third_party.containers.rlocations import REGISTRY_2_TARBALL, RYUK_TARBALL
-from util.oci import BazelImage, load_image
+from util.crane import BazelImage
+from util.oci import load_image
 
 logger = logging.getLogger(__name__)
 tracer = trace.get_tracer(__name__)
